requirements.txt
torch>=1.7.0
torchvision>=0.8.0
opencv-python>=4.5.0
numpy>=1.21.0
pycocotools>=2.0.0
albumentations
pycocotools
tqdm>=4.62.0
cython
scipy>=1.7.0
motmetrics
opencv-python
seaborn
lap
Pillow>=8.3.0
matplotlib>=3.4.0
PyYAML>=5.4.0
tensorboard>=2.7.0